+++ Source/Scene/ClassificationPrimitive.js
@@ -6,7 +6,6 @@ import defineProperties from '../Core/defineProperties.js';
 import destroyObject from '../Core/destroyObject.js';
 import DeveloperError from '../Core/DeveloperError.js';
 import GeometryInstance from '../Core/GeometryInstance.js';
-import isArray from '../Core/isArray.js';
 import DrawCommand from '../Renderer/DrawCommand.js';
 import Pass from '../Renderer/Pass.js';
 import RenderState from '../Renderer/RenderState.js';
@@ -173,7 +172,7 @@ import StencilOperation from './StencilOperation.js';
         this.appearance = options.appearance;
 
         var readOnlyAttributes;
-        if (defined(geometryInstances) && isArray(geometryInstances) && geometryInstances.length > 1) {
+        if (defined(geometryInstances) && Array.isArray(geometryInstances) && geometryInstances.length > 1) {
             readOnlyAttributes = ClassificationPrimitiveReadOnlyInstanceAttributes;
         }
 
@@ -1002,7 +1001,7 @@ import StencilOperation from './StencilOperation.js';
         var primitiveOptions = this._primitiveOptions;
 
         if (!defined(this._primitive)) {
-            var instances = isArray(this.geometryInstances) ? this.geometryInstances : [this.geometryInstances];
+            var instances = Array.isArray(this.geometryInstances) ? this.geometryInstances : [this.geometryInstances];
             var length = instances.length;
 
             var i;