+++ Source/Widgets/Viewer/Viewer.js
@@ -10,7 +10,6 @@ import DeveloperError from '../../Core/DeveloperError.js';
 import Event from '../../Core/Event.js';
 import EventHelper from '../../Core/EventHelper.js';
 import HeadingPitchRange from '../../Core/HeadingPitchRange.js';
-import isArray from '../../Core/isArray.js';
 import Matrix4 from '../../Core/Matrix4.js';
 import ScreenSpaceEventType from '../../Core/ScreenSpaceEventType.js';
 import BoundingSphereState from '../../DataSources/BoundingSphereState.js';
@@ -444,7 +443,7 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
             toolbar.appendChild(geocoderContainer);
             var geocoderService;
             if (defined(options.geocoder) && typeof options.geocoder !== 'boolean') {
-                geocoderService = isArray(options.geocoder) ? options.geocoder : [options.geocoder];
+                geocoderService = Array.isArray(options.geocoder) ? options.geocoder : [options.geocoder];
             }
             geocoder = new Geocoder({
                 container : geocoderContainer,
@@ -1820,7 +1819,7 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
             }
 
             //Zoom target is already an array, just copy it and return.
-            if (isArray(zoomTarget)) {
+            if (Array.isArray(zoomTarget)) {
                 that._zoomTarget = zoomTarget.slice(0);
                 return;
             }
@@ -1834,7 +1833,7 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
             }
 
             //Zoom target is already an array, just copy it and return.
-            if (isArray(zoomTarget)) {
+            if (Array.isArray(zoomTarget)) {
                 that._zoomTarget = zoomTarget.slice(0);
             } else {
                 //Single entity