+++ Source/Workers/cesiumWorkerBootstrapper.js
@@ -63,10 +63,6 @@ var requirejs, require, define;
         return ostring.call(it) === '[object Function]';
     }
 
-    function isArray(it) {
-        return ostring.call(it) === '[object Array]';
-    }
-
     /**
      * @private
      * Helper function for iterating over an array. If the func returns
@@ -134,7 +130,7 @@ var requirejs, require, define;
             eachProp(source, function (value, prop) {
                 if (force || !hasProp(target, prop)) {
                     if (deepStringMixin && typeof value === 'object' && value &&
-                        !isArray(value) && !isFunction(value) &&
+                        !Array.isArray(value) && !isFunction(value) &&
                         !(value instanceof RegExp)) {
 
                         if (!target[prop]) {
@@ -395,7 +391,7 @@ var requirejs, require, define;
 
         function hasPathFallback(id) {
             var pathConfig = getOwn(config.paths, id);
-            if (pathConfig && isArray(pathConfig) && pathConfig.length > 1) {
+            if (pathConfig && Array.isArray(pathConfig) && pathConfig.length > 1) {
                 //Pop off the first array value, since it failed, and
                 //retry
                 pathConfig.shift();
@@ -1345,7 +1341,7 @@ var requirejs, require, define;
                 if (cfg.shim) {
                     eachProp(cfg.shim, function (value, id) {
                         //Normalize the structure
-                        if (isArray(value)) {
+                        if (Array.isArray(value)) {
                             value = {
                                 deps: value
                             };
@@ -1678,7 +1674,7 @@ var requirejs, require, define;
                         if (parentPath) {
                             //If an array, it means there are a few choices,
                             //Choose the one that is desired
-                            if (isArray(parentPath)) {
+                            if (Array.isArray(parentPath)) {
                                 parentPath = parentPath[0];
                             }
                             syms.splice(0, i, parentPath);
@@ -1774,10 +1770,10 @@ var requirejs, require, define;
             contextName = defContextName;
 
         // Determine if have config object in the call.
-        if (!isArray(deps) && typeof deps !== 'string') {
+        if (!Array.isArray(deps) && typeof deps !== 'string') {
             // deps is a config object
             config = deps;
-            if (isArray(callback)) {
+            if (Array.isArray(callback)) {
                 // Adjust args if there are dependencies
                 deps = callback;
                 callback = errback;
@@ -2066,7 +2062,7 @@ var requirejs, require, define;
         }
 
         //This module may not have dependencies
-        if (!isArray(deps)) {
+        if (!Array.isArray(deps)) {
             callback = deps;
             deps = null;
         }