+++ CHANGES.md
@@ -5,6 +5,7 @@ Change Log
 
 ##### Deprecated :hourglass_flowing_sand:
 * The property `Scene.sunColor` has been deprecated and will be removed in Cesium 1.69. Use `scene.light.color` and `scene.light.intensity` instead. [#8493](https://github.com/AnalyticalGraphicsInc/cesium/pull/8493)
+* The `isArray` function has been deprecated and will be removed in Cesium 1.69. Use the native `Array.isArray` function instead. [#8526](https://github.com/AnalyticalGraphicsInc/cesium/pull/8526)
 
 ##### Additions :tada: